Add tycho-emma-coverage-plugin
Add EmmaAutomationMojo which uses tycho.testArgLine to automatically configure
the
bundle to instrument and to exclude the test classes from instrumentation.
For details and usage see included integration test.
Adding org.eclemma.runtime.equinox bundle as resource for the integration test
here since it is not available in a public maven repo. See
https://docs.sonatype.org/display/TYCHO/How+to+integrate+EMMA+%28code+coverage%29+with+Tycho
which explains how to repackage the offical eclemma.zip.
Discussion with Dave Meibusch how to update the wiki-page pending.