Run emma ITs as part of the main build
Move emma ITs into module tycho-emma-its
to keep related code together.
Collapsing ITs into module tycho-emma-coverage-plugin
was not possible because integration tests are forking
a separate maven process which expects the artifact
under test (which is just being built but not installed yet)
to be already available in the local repo.