Mirror mojo: mirror specific IUs, expose slicing options
We can now mirror specific IUs only and
configure various slicing options.
This makes MirrorMojo an equivalent replacement of
the p2.mirror ant task [1].

In particular, the plan is to eventually use this mojo for
mirroring the target platform of
the tycho ITs so we don't have to require
a local eclipse installation for executing
tycho ITs anymore.

[1] http://help.eclipse.org/indigo/index.jsp?topic=%2Forg.eclipse.platform.doc.isv%2Fguide%2Fp2_repositorytasks.htm